promql: fix unconvert issues (#4040)
See,
$ gometalinter --vendor --disable-all --enable=unconvert --deadline 6m ./...
promql/engine.go:1396:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1396:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1398:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1398:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1427:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1427:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1429:26:warning: unnecessary conversion (unconvert)
promql/engine.go:1429:40:warning: unnecessary conversion (unconvert)
promql/engine.go:1505:50:warning: unnecessary conversion (unconvert)
promql/engine.go:1573:46:warning: unnecessary conversion (unconvert)
promql/engine.go:1578:46:warning: unnecessary conversion (unconvert)
promql/engine.go:1591:80:warning: unnecessary conversion (unconvert)
promql/engine.go:1602:94:warning: unnecessary conversion (unconvert)
promql/engine.go:1630:18:warning: unnecessary conversion (unconvert)
promql/engine.go:1631:24:warning: unnecessary conversion (unconvert)
promql/engine.go:1634:18:warning: unnecessary conversion (unconvert)
promql/engine.go:1635:34:warning: unnecessary conversion (unconvert)
promql/functions.go:302:42:warning: unnecessary conversion (unconvert)
promql/functions.go:315:42:warning: unnecessary conversion (unconvert)
promql/functions.go:334:26:warning: unnecessary conversion (unconvert)
promql/functions.go:395:31:warning: unnecessary conversion (unconvert)
promql/functions.go:406:31:warning: unnecessary conversion (unconvert)
promql/functions.go:454:27:warning: unnecessary conversion (unconvert)
promql/functions.go:701:46:warning: unnecessary conversion (unconvert)
promql/functions.go:701:78:warning: unnecessary conversion (unconvert)
promql/functions.go:730:43:warning: unnecessary conversion (unconvert)
promql/functions.go:1220:23:warning: unnecessary conversion (unconvert)
promql/functions.go:1249:23:warning: unnecessary conversion (unconvert)
promql/quantile.go:107:54:warning: unnecessary conversion (unconvert)
promql/quantile.go:182:16:warning: unnecessary conversion (unconvert)
promql/quantile.go:182:64:warning: unnecessary conversion (unconvert)

+++ promql/engine.go
@@ -1393,9 +1393,9 @@ func scalarBinop(op ItemType, lhs, rhs float64) float64 {
 	case itemDIV:
 		return lhs / rhs
 	case itemPOW:
-		return math.Pow(float64(lhs), float64(rhs))
+		return math.Pow(lhs, rhs)
 	case itemMOD:
-		return math.Mod(float64(lhs), float64(rhs))
+		return math.Mod(lhs, rhs)
 	case itemEQL:
 		return btos(lhs == rhs)
 	case itemNEQ:
@@ -1424,9 +1424,9 @@ func vectorElemBinop(op ItemType, lhs, rhs float64) (float64, bool) {
 	case itemDIV:
 		return lhs / rhs, true
 	case itemPOW:
-		return math.Pow(float64(lhs), float64(rhs)), true
+		return math.Pow(lhs, rhs), true
 	case itemMOD:
-		return math.Mod(float64(lhs), float64(rhs)), true
+		return math.Mod(lhs, rhs), true
 	case itemEQL:
 		return lhs, lhs == rhs
 	case itemNEQ:
@@ -1502,7 +1502,7 @@ func (ev *evaluator) aggregation(op ItemType, grouping []string, without bool, p
 			lb.Del(labels.MetricName)
 		}
 		if op == itemCountValues {
-			lb.Set(valueLabel, strconv.FormatFloat(float64(s.V), 'f', -1, 64))
+			lb.Set(valueLabel, strconv.FormatFloat(s.V, 'f', -1, 64))
 		}
 
 		var (
@@ -1570,12 +1570,12 @@ func (ev *evaluator) aggregation(op ItemType, grouping []string, without bool, p
 			group.groupCount++
 
 		case itemMax:
-			if group.value < s.V || math.IsNaN(float64(group.value)) {
+			if group.value < s.V || math.IsNaN(group.value) {
 				group.value = s.V
 			}
 
 		case itemMin:
-			if group.value > s.V || math.IsNaN(float64(group.value)) {
+			if group.value > s.V || math.IsNaN(group.value) {
 				group.value = s.V
 			}
 
@@ -1588,7 +1588,7 @@ func (ev *evaluator) aggregation(op ItemType, grouping []string, without bool, p
 			group.groupCount++
 
 		case itemTopK:
-			if int64(len(group.heap)) < k || group.heap[0].V < s.V || math.IsNaN(float64(group.heap[0].V)) {
+			if int64(len(group.heap)) < k || group.heap[0].V < s.V || math.IsNaN(group.heap[0].V) {
 				if int64(len(group.heap)) == k {
 					heap.Pop(&group.heap)
 				}
@@ -1599,7 +1599,7 @@ func (ev *evaluator) aggregation(op ItemType, grouping []string, without bool, p
 			}
 
 		case itemBottomK:
-			if int64(len(group.reverseHeap)) < k || group.reverseHeap[0].V > s.V || math.IsNaN(float64(group.reverseHeap[0].V)) {
+			if int64(len(group.reverseHeap)) < k || group.reverseHeap[0].V > s.V || math.IsNaN(group.reverseHeap[0].V) {
 				if int64(len(group.reverseHeap)) == k {
 					heap.Pop(&group.reverseHeap)
 				}
@@ -1627,12 +1627,12 @@ func (ev *evaluator) aggregation(op ItemType, grouping []string, without bool, p
 			aggr.value = float64(aggr.groupCount)
 
 		case itemStdvar:
-			avg := float64(aggr.value) / float64(aggr.groupCount)
-			aggr.value = float64(aggr.valuesSquaredSum)/float64(aggr.groupCount) - avg*avg
+			avg := aggr.value / float64(aggr.groupCount)
+			aggr.value = aggr.valuesSquaredSum/float64(aggr.groupCount) - avg*avg
 
 		case itemStddev:
-			avg := float64(aggr.value) / float64(aggr.groupCount)
-			aggr.value = math.Sqrt(float64(aggr.valuesSquaredSum)/float64(aggr.groupCount) - avg*avg)
+			avg := aggr.value / float64(aggr.groupCount)
+			aggr.value = math.Sqrt(aggr.valuesSquaredSum/float64(aggr.groupCount) - avg*avg)
 
 		case itemTopK:
 			// The heap keeps the lowest value on top, so reverse it.
